add pom, cat test, jacoco, Lion Test
