+++ src/test/java/com/example/LionUsesParameterHasManeTest.java
@@ -1,24 +1,20 @@
 package com.example;
 
 import junit.framework.TestCase;
-import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.junit.runners.Parameterized;
-
-
 
 @RunWith(Parameterized.class)
 public class LionUsesParameterHasManeTest extends TestCase {
 
     private Feline feline;
     private final String sexLion;
-    private final boolean expectedHasMane;
+    private final boolean HasMane;
 
-
-    public LionUsesParameterHasManeTest(String sexLion, boolean expectedHasMane) {
+    public LionUsesParameterHasManeTest(String sexLion, boolean HasMane) {
         this.sexLion = sexLion;
-        this.expectedHasMane = expectedHasMane;
+        this.HasMane = HasMane;
     }
 
     @Parameterized.Parameters
@@ -26,22 +22,19 @@ public class LionUsesParameterHasManeTest extends TestCase {
         return new Object[][] {
                 {"Самец", true},
                 {"Самка", false},
-                {"Мужской", true},
-                {"Женский", false},
-        };
+
+                 };
     }
 
     @Test
     public void shouldDoesHaveMane() throws Exception {
-       Lion lion = new Lion(this.sexLion,this.feline);
+        Lion lion = new Lion(this.sexLion,this.feline);
+        boolean expectedHasMane =this.HasMane;
 
-       boolean actualHasMane = lion.doesHaveMane();
+        boolean actualHasMane = lion.doesHaveMane();
 
-       assertEquals("Некорректное соотношение гривы и пола животного", this.expectedHasMane, actualHasMane);
+        assertEquals("Некорректное соотношение гривы и пола животного", expectedHasMane, actualHasMane);
     }
-
-
-
 
 
 }
